Split table operations to more files
